feat: Implemented JobRunner, TestRegistry, Reporter
JobRunner is orchestrator that manage concurrently ran tests.It iterates over tests provided from user, perfmorms them and adds to the shared channel. It is basic implementation of this module and will be develop later.

TestRegistry is module that initializes available tests in the program. .It helps to statelessly run tests.

Reporter module is obliged to get the result from shared channel and send it to backend. Request sending logic will be implemented later.